refactor: 테스트 Fixture 분리, 네이밍 수정
