+++ src/test/java/roomescape/DateUtils.java
@@ -1,0 +1,18 @@
+package roomescape;
+
+import java.time.LocalDate;
+
+public class DateUtils {
+
+    public static LocalDate yesterday() {
+        return LocalDate.now().minusDays(1);
+    }
+
+    public static LocalDate today() {
+        return LocalDate.now();
+    }
+
+    public static LocalDate tomorrow() {
+        return LocalDate.now().plusDays(1);
+    }
+}
